requirements_20250227.txt
aggdraw==1.3.16
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.3
cachetools==5.5.2
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==3.1.0
click==8.1.3
clr_loader==0.2.7.post0
colorama==0.4.6
Deprecated==1.2.14
distro==1.9.0
docopt==0.6.2
eng-to-ipa==0.0.2
fastapi==0.112.0
frozenlist==1.3.3
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.3.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
googletrans-python==0.2.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.5.0
httpx==0.28.1
hyperframe==6.0.1
idna==3.4
imageio==2.31.1
iso8601==2.1.0
jaconv==0.3.4
Janome==0.4.2
jiter==0.8.0
lazy_loader==0.3
Levenshtein==0.25.0
lxml==5.3.0
lxml-stubs==0.5.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
ndgr-client @ git+https://github.com/tsukumijima/NDGRClient@ccd2eed16f6f6d3b77fa7ed2be0ffce0c3999cb5
networkx==3.1
numpy==1.24.2
openai==1.55.3
packaging==23.1
pefile==2023.2.7
Pillow==9.4.0
proto-plus==1.26.0
protobuf==5.27.3
psd-tools==1.9.28
psutil==5.9.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.13
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
pykakasi==2.2.1
pyparsing==3.2.1
pytchat==0.5.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.9
pythonnet==3.0.5
PyWavelets==1.4.1
pywin32==305
pywin32-ctypes==0.2.2
PyYAML==6.0
rapidfuzz==3.7.0
requests==2.28.2
rich==13.7.1
rsa==4.9
scikit-image==0.21.0
scipy==1.11.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
starlette==0.37.2
tifffile==2023.7.18
tqdm==4.65.0
twitchAPI==4.2.1
twitchio==2.10.0
typer==0.12.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.21.1
watchfiles==0.18.1
websocket-client==1.7.0
websockets==14.2
wrapt==1.16.0
yarl==1.8.2
